requirements.txt
PyYAML~=6.0
Pillow~=8.4.0
